requirements.txt
Bottleneck==1.4.2
FXrays==1.3.6
GitPython==3.1.45
ImageIO==2.37.2
Jinja2==3.1.6
MarkupSafe==3.0.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
PySocks==1.7.1
PyX==0.17
PyYAML==6.0.3
Pygments==2.19.1
ace_tools_open==0.1.0
altair==5.5.0
altgraph==0.17.5
annotated-types==0.7.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotlicffi==1.1.0.0
bs4==0.0.2
build==1.3.0
cachetools==6.2.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
clifford==1.5.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
cypari==2.5.6
decorator==5.2.1
dill==0.4.0
docutils==0.22.3
dpcpp-cpp-rt==2025.0.4
executing==2.2.1
filelock==3.24.3
fonttools==4.56.0
freetype-py==2.5.1
fsspec==2026.2.0
gitdb==4.0.12
h5py==3.15.1
hsluv==5.0.4
id==1.5.0
idna==3.11
ijson==3.5.0
iniconfig==2.0.0
intel-cmplr-lib-rt==2025.0.4
intel-cmplr-lib-ur==2025.0.4
intel-cmplr-lic-rt==2025.0.4
intel-opencl-rt==2025.0.4
intel-openmp==2025.0.4
intel-sycl-rt==2025.0.4
intel_extension_for_pytorch==2.5.10+xpu
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itables==2.7.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jsonschema==4.25.0
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.4.8
knot_floer_homology==1.2.2
lazy_loader==0.4
llvmlite==0.46.0
low_index==1.2.1
markdown-it-py==3.0.0
matplotlib==3.10.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mkl==2025.0.1
mkl-dpcpp==2025.0.1
mkl-service==2.5.2
mkl_fft==2.1.1
mkl_random==1.3.0
more-itertools==10.8.0
mpl-toolkits.clifford==0.0.3
mpmath==1.3.0
narwhals==2.13.0
networkx==3.5
nh3==0.3.2
numba==0.63.0
numexpr==2.14.1
numpy==2.2.4
onemkl-sycl-blas==2025.0.1
onemkl-sycl-datafitting==2025.0.1
onemkl-sycl-dft==2025.0.1
onemkl-sycl-lapack==2025.0.1
onemkl-sycl-rng==2025.0.1
onemkl-sycl-sparse==2025.0.1
onemkl-sycl-stats==2025.0.1
onemkl-sycl-vm==2025.0.1
openvino==2025.2.0
openvino-genai==2025.2.0.0
openvino-telemetry==2025.2.0
openvino-tokenizers==2025.2.0.1
packaging==24.2
pandas==2.3.3
pandas-stubs==2.3.3.251219
parso==0.8.6
pefile==2024.8.26
pickleshare==0.7.5
pillow==11.1.0
plink==2.4.9
plotly==6.5.0
pluggy==1.5.0
prompt_toolkit==3.0.52
protobuf==5.29.3
psutil==7.2.2
pure_eval==0.2.3
pyarrow==21.0.0
pybind11==2.13.6
pybind11_global==2.13.6
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
pyparsing==3.2.1
pypng==0.20220715.0
pyproject_hooks==1.2.0
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32-ctypes==0.2.3
readme_renderer==44.0
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rpds-py==0.28.0
ruamel.yaml==0.19.1
scikit-image==0.25.2
scipy==1.16.1
scipy-stubs==1.16.3.2
setuptools==80.9.0
shiboken6==6.10.2
six==1.17.0
smmap==5.0.2
snappy==3.3
snappy_manifolds==1.4
soupsieve==2.8.3
sparse==0.17.0
spherogram==2.4.1
stack-data==0.6.3
streamlit==1.52.0
sympy==1.13.1
tbb==2022.3.1
tcmlib==1.2.0
tenacity==9.1.2
tifffile==2025.10.4
tkinter-gl==1.1
toml==0.10.2
torch==2.5.1+cxx11.abi
torchaudio==2.5.1+cxx11.abi
torchvision==0.20.1+cxx11.abi
tornado==6.5.1
tqdm==4.67.3
traitlets==5.14.3
trimesh==4.10.1
twine==6.2.0
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
umf==0.9.1
urllib3==2.5.0
vispy==0.16.1
watchdog==6.0.0
wcwidth==0.6.0
wheel==0.45.1
win_inet_pton==1.1.0